[BE] feat : 공공 Api Json 데이터 받기를 담당하는 클래스 생성
- OpenApiUtils.java : 중복을 제거하기 귀해 공공 Api  Json 데이터 받기를 담당하는 클래스 생성했습니다.

associate with #7
